refactor cli runtime orchestration
