+++ src/lib.rs
@@ -16,6 +16,7 @@ macro_rules! log_status {
 
 extern crate self as homeboy;
 
+pub mod cli_runtime;
 pub mod cli_surface;
 pub mod command_contract;
 #[doc(hidden)]
